Heat client fixture needs to support a parameter for expected target stack status
Issues:
Fixes #40

Problem:
In order to extend these fixtures to work with the
f5-openstack-heat-plugins functional tests, we need to expect a stack to
fail. One way we do this is to create a stack and expect the status to
be 'CREATED_FAILED'. So we will allow a new parameter when creating a
stack which is the expect target status.

Analysis:
Added the default target status to 'CREATE_COMPLETE' and allowed a
consumer to pass in their expected status.

Tests:
Ran tests against the f5-openstack-heat-plugins repo with success